Update README.md
E-pay-payment-system started as an idea aimed to help in payment of school fees in Kenya after I noticed that most students have to wait in long queues in banking halls in order to make the payments. The aim of E-pay was to solve this problem. The other problem that I noticed that needed a solution was that many people did not have access to smartphones at the time and many rural areas also experienced internet connection problems. To solve this issue, I wrote Python code that could enable users to utilize USSD code at a low charge to complete their payments.